+++ .env.development
@@ -1,1 +1,1 @@
-ESLINT_NO_DEV_ERRORS=true
+ESLINT_NO_DEV_ERRORS=false
